feature: Take auth credentials from a Secret, closes #18
The GatewayConfig can contain an optional AuthRef field, which must be a proper reference to a
Secret resource containing the TURN authentication credentials. The following rules apply

AuthRef holds an optional reference to a Secret that specifies the TURN authentication
credentials for STUNner.  The following conditions must hold:
- group MUST be set to "" (corev1.GroupName), "v1", or omitted,
- kind MUST be set to "Secret" or omitted,
- name MUST be the name of a valid Secret,
- namespace MAY be omitted, in which case it defaults to the namespace of
  the GatewayConfig, or it MAY be any valid namespace where the Secret lives.

The referenced Secret MUST be of type Opaque and the following conditions MUST hold:
- the Secret MUST contain a "type" field that MUST be set to either
  "plaintext" or "longterm" (but see
  https://github.com/l7mp/stunner-gateway-operator/issues/7),
- if type is "plaintext" then the Secret MUST contain a "username" and a
  "password" field that together specify the username/password pair clients can use to
  authenticate with Stunner,
- if type is "lonterm" then the Secret MUST contain a single field named
  "sharedSecret" or "secret" that is used to check the authenticity of time-windowed
  TURN credentials.

Note that externally set credentials override any inline auth credentials (AuthType,
AuthUsername, etc.): if AuthRef is nonempty then it is expected that the referenced
Secret exists and *all* authentication credentials are correctly set in the referenced
Secret (username/password or shared secret). Mixing of credential sources
(inline/external) is not supported.