fix: LB service watchers now filter on the label "app:stunner"
So far we have filtered on the "related-gateway" annotation but since in Kubernetes server-side
filtering is not available on annotations, only labels, we had to list *all* services from
Kubernetes and filter locally in the controller. With this change, we can now filter at the server
side, which will improve scalability to very large clusters.