feature: Introduce auth type aliases, fixes #7
STUNner' current terminology is doubly confusing: (1) what we call plaintext is not plain-text at
all, and (2) we reuse the name longterm from the STUN spec to mean the time-windowed authentication
mode but in fact both what we call plaintext and longterm use the STUN "long-term" authentication
mechanism.

This commit introduces the following aliases in the GatewayConfig authType field: (1) static as an
alias for plaintext, and (2) timewindowed and ephemeral as aliases for longterm.

For backward compatibility, the dataplane API does not change.