feat: Implement a StaticService CRD
StaticService is a simple Service that wraps a static set of IP prefixes and can stand as a backend
when we want to access to a set of known IP prefixes through STUNner. This is particularly useful
when STUNner is used as a standalone public TURN server, where there is no Kubernetes service that
would abstract the user base which is now outside the cluster.