Feature: managed dataplane (#35)
Major feature addition to allow the operator to manage the stunnerd pods per each Gateway. Apart from managing a LB Service and a ConfigMap holding the running dataplane configuration, the operator can now also manage the stunnerd Deployment for Gateways. This makes it possible to deploy STUNner from a single Helm chart and simplifies dealing with multiple stunnerd versions inside the same cluster (e.g., stable and dev). Legacy mode is still available.

- Managed mode can be enabled by starting the operator with the command line argument `-dataplane-mode=managed` (default is off). 
- Introduce a new `Dataplane` cluster-scoped CRD that defines the stunnerd templates (container image, command line arguments, etc.) used to spawn stunnerd pods per Gateways (see sample in `deploy/manifests/default_dataplane.yaml`). The Dataplane for a gateway hierarchy can be specified in the GatewayConfig; if no dataplane is defined then the `default` Dataplane is used.
- Major refactoring in the config renderer code to allow the same renderer to generate the config per each GatewayClass (legacy mode) or per each Gateway (managed mode).
- More consistent labels/annotations on managed resources. This affects legacy dataplane mode as well.
- Operator-sdk version, envtest and Go pkg version updates all over the place.

Linter checks fail due to the Ready status becoming deprecated in the new Gateway API version we are using, this will be taken care of in a subsequent PR.