feature: Implement Dataplane templates
It is now possible to select the template used by the operator deploy the dataplane pods in the
managed dataplane mode. The template can be selected in the Dataplane CRD (`spec.template`).
Currently there are two templates implemented. The `default` template spawns a single stunnerd
container with the running config mapped into the pod from a ConfigMap. The `config-watcher`
template uses a separate sidecar container to watch the ConfigMap, which can be faster than the
default but requires additional RBAC permissions.