feat: Implement a Config Discovery Service (CDS) client
The CDS service allows the STUNner dataplane component `stunnerd` to autodiscover its own
configuration. This is achieved by `stunnerd` using a WebSocket connection to the CDS server (this
operator) to request its configuration. As long as the connection is open, the CDS server will push
new configurations down to the client as they become available. The config file watcher also
remains available for the "legacy" dataplane mode.

- implement a CDS server and integrate it into the startup sequence
- managed dataplane mode now switches on the CDS server by default
- remove "template" from the Dataplane spec: the purpose of this was to allow the operator to spanw
  a config-watcher sidecar container for stunnerd but since this is no longer needed the template
  has gone
- heath-check port and metrics scraping port are now configurable via the Dataplane CRD
  (overwritten by the settings in the GatewayConfig spec)
- the dataplane Deployment now correctly sets the health-check port in managed mode
- inject name/namespace in env variables as well as the CDS server URL to stunnerd in managed mode
- CDS tests