fix: Make sure service-port names of LB Services are unique, fixes #19
Service-port names used to be created from the Gateway name plus listener port, which led to a name clash for Gateways containing multiple listeners for the same protocol. This commit changes this behavior in that service-port names are now created from the listener name, which is supposed to be unique per Gateway.