squash! fix: Nonzero replica-count in Dataplane.spec no longer prevents autoscaling
fix: No longer set spec.externalIPs on LB Services

So far we have set both spec.LoadBalancerIP and spec.externalIPs on the LB Service created for a
gateway based on gateway.spec.addresses. This was intended to be on the safe side, as
spec.LoadBalancerIP is deprecated. However, this seems to break GKE. This commit removes the code
setting spec.externalIPs: this field serves a completely different purpose and this feature was
never tested on a real cloud anyway.

# Conflicts:
#	internal/renderer/service_util.go