fix: Remove dataplane ConfigMap, fixes #43
So far we have kept around a STUNner dataplane ConfigMap per each Gateway holding the running
dataplane config for the Gateway. The ConfigMap was automatically created by the operator even
though in managed mode we no longer use it to configure the stunnerd pods (this is now done via the
Config Discovery Service), since many STUNner components and utilities (auth-service, stunnerctl,
turncat) still use it to read dataplane configs. Now that all STUNner components have been upgraded
to CDS this ConfigMap is no longer used, so this commit removes it.