feature: Allow for disabling the managed dataplane for a Gateway
In some cases it may be useful to manually provision a dataplane for a Gateway, e.g., to deploy
`stunnerd` in a DeamonSet instead of a Deployment. This commit makes this possible: add an
annotation `stunner.l7mp.io/disable-managed-dataplane:true` to a Gateway to prevent STUNner from
spawning a dataplane for it.