fix: Exit when the operator fails to start up
We forgot to add an exit call to the `operator. Start()` call's error path. This leads to strange bugs then the operator cannot access the API server due to a transient problem, the operator fails to register itself as a Gateway API controller so it will not get the watch events it should handle, but instead of exiting it's just sitting there doing nothing. This change makes sure the operator restarts when the API server is unreachable.